+++ polardb-20170801/core/Models/RegisterKnowledgeBaseFileRequest.cs
@@ -1,0 +1,47 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Polardb20170801.Models
+{
+    public class RegisterKnowledgeBaseFileRequest : TeaModel {
+        /// <summary>
+        /// <para>The OSS object key of the uploaded file, excluding the oss://BucketName/ prefix.</para>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>pks-2ze123456789abcd/pkb-2ze123456789abcd/example.pdf</para>
+        /// </summary>
+        [NameInMap("FilePath")]
+        [Validation(Required=false)]
+        public string FilePath { get; set; }
+
+        /// <summary>
+        /// <para>The knowledge base ID.</para>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>pkb-2ze123456789abcd</para>
+        /// </summary>
+        [NameInMap("KnowledgeBaseId")]
+        [Validation(Required=false)]
+        public string KnowledgeBaseId { get; set; }
+
+        /// <summary>
+        /// <para>The ID of the region where the knowledge base resides.</para>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>cn-beijing</para>
+        /// </summary>
+        [NameInMap("RegionId")]
+        [Validation(Required=false)]
+        public string RegionId { get; set; }
+
+    }
+
+}